[TESTS] Move base58 to own module to break circular dependency
This breaks the script->key->address->script dependency cycle.
