+++ test/functional/test_runner.py
@@ -67,7 +67,7 @@ TEST_EXIT_PASSED = 0
 TEST_EXIT_SKIPPED = 77
 
 TEST_FRAMEWORK_MODULES = [
-    "address",
+    "base58",
     "blocktools",
     "script",
 ]
